refactor: move the __repr__ method to resource representation helper
